+++ packages/preconfirmation-p2p/crates/net/src/driver/mod.rs
@@ -111,10 +111,8 @@ impl NetworkDriver {
         validator: Box<dyn ValidationAdapter>,
         storage: Option<Arc<dyn PreconfStorage>>,
     ) -> anyhow::Result<(Self, NetworkHandle)> {
-        let dial_factor = {
-            let (_, _, _, _, _, _, dial) = cfg.resolve_connection_caps();
-            NonZeroU8::new(dial).unwrap_or_else(|| NonZeroU8::new(1).unwrap())
-        };
+        let dial_factor =
+            NonZeroU8::new(cfg.dial_concurrency_factor).unwrap_or(NonZeroU8::new(1).unwrap());
 
         let parts = build_transport_and_behaviour(&cfg)?;
         let peer_id = parts.keypair.public().to_peer_id();
