--- packages/taiko-client-rs/crates/driver/src/signer/mod.rs
@@ -1,3 +1,0 @@
-pub mod fixed_k;
-
-pub use fixed_k::{FixedKSigner, FixedKSignerError, SignatureWithRecoveryId};
